CA-400124: rrd: Serialize transform parameter for data sources
Previously, the transform function was not serialized in the plugin-server
protocol (it was only used in xcp_rrdd itself, not in plugins). This issue was
revealed by the previous work around splitting cpu metrics into a separate plugin.

Instead of allowing arbitrary functions (which would be difficult to
serialize), for 'fun x' offer just two options:
* Inverse (1.0 - x), and
* Identity (x)

Default (if the parameter is not provided, either in OCaml or JSON), is
Identity.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>