CP-51585: Add knowledge of baggage to http_lib
Provides facilities to associate trace baggage with HTTP requests.

The provided functions are akin to the "inject" and "extract" functions
from trace propagator literature. The difference is that, currently, we
have traceparent as a first-class member within a HTTP request (which
gets transmitted as HTTP header), so code that creates derived spans
from requests (with tracing) query for this member to create a derived
span, as opposed to extracting out an entire SpanContext from a request
(which is the general approach with inject/extract function pairs; they
generally manipulate information at the granularity of an entire span
context - to keep everything together).

In future, we can:

- Remove "traceparent" as an explicit member within the record for a
  HTTP request. Tracing related information would be propagated inside
  the additional_headers field.
- Manipulate tracing-related context using an entire SpanContext. This
  helps to avoid having a lot of ad-hoc logic for dealing with each
  potential propagator encoding with the carrier (the verbosity of this
  is exacerbated by the opt-in nature of the tracing information).
- Ensure "baggage" is only propagated in the way we intend, it generally
  should not be caught by current logic which encodes extraneous headers
  as tracing attributes of a specific format.
- According to the standard, special header entries like "baggage" can
  actually occur multiple times within a HTTP header. Currently, we use
  List.assoc(_opt), but - in future - we may wish to write a function:
  ```
    val all_assoc : 'a -> ('a * 'b) list -> 'b list
  ```
  to extract all entries, and then use the combination mechanism
  provided in this commit to fold them together (preferring the later
  entries). That would make us more compliant with standards.

Signed-off-by: Colin James <colin.barr@cloud.com>