CP-51585: Add baggage as attributes during export
Modifies exportation of spans to the Zipkinv2 format to reify baggage as
attributes. In future, we could be more selective about where we choose
to export baggage as attributes: there's probably enough querying logic
in Jaeger etc. to allow us to easily group tagged spans with their
descendents.

If we didn't commit the baggage into span content in some way, we may
never see it. Baggage is inter-service metadata - meaning
applications can do whatever they want with it (such as ignore it).

Signed-off-by: Colin James <colin.barr@cloud.com>