CP-51585: Recreate span context in HTTP tracing
Currently, when you rewrite a HTTP request to have tracing, using
Http.Request.with_tracing, it only considers the traceparent in order to
construct a SpanContext.

This change makes sure that the span context given to the child span
contains the baggage information encoded in the request. The baggage
itself is propagated at the protocol level (as it exists within
additional_headers), but not at the service level (tracing level).

Signed-off-by: Colin James <colin.barr@cloud.com>