CP-51585: Factor out common baggage pattern
Currently, propagation of baggage only works if a traceparent is
provided. This is because the tracing boundaries were initially written
only to support traceparent propagation.

Therefore, code that applies baggage is contained within continuations
of monadic binds applied to this (optional) traceparent value: making it
dependent upon a traceparent being provided. This was convenient,
initially, as it made it obvious where boundary trace propagation
happens (you just search for traceparent and see it being extracted
out). However, this design is problematic as baggage is applied to the
SpanContext and, if no traceparent is provided, there is no way to push
that down (as the carrier context is not pushed down the call stack far
enough to be useful).

This change seeks to remove the same design pattern from being
implemented ad-hoc at every usage site: optionally folding baggage into
a span context is now done with a "with_baggage_maybe" function.

This change should be temporary: the next phase to make baggage
independent of traceparent is to encapsulate both of them into a
serialisable "boundary context" which can be pushed down at certain
sites.

Signed-off-by: Colin James <colin.barr@cloud.com>