Replace fold with of_list in rbac
Folding over a list to add its elements to a set (which is initially empty) is
operationally equivalent to calling of_list (of the set), but
potentially less efficient.

The implementation of of_list only uses "add" for small lists, e.g. the
cases for lists [x_1; x_2; ...; x_N] for all N in range 2 <= N <= 5
are matched literally and expanded to:

   add x_N (... (add x_1 (singleton x_0)))

However, larger lists are first sorted and the underlying tree
representing the set is constructed directly.

Signed-off-by: Colin James <colin.barr@cloud.com>