CP-51772: Consolidate propagation into tracing lib
Generalise the pattern used in http_svr.ml into a functor in the tracing
library. In particular, so long as you can provide a way to name a
carrier (to endow its child trace with), inject tracing context into a
carrier, and extract trace context into a carrier, you can use this
generic pattern to propagate tracing across arbitrary carriers (where
the derived span is exported, but the incoming one is not).

To this end, we factor it out of http_svr.ml and redefine the helper
module used there in terms of one constructed from the new tracing
propagator functor.

The tracing_propagator library is used to provide definitions in the
input module for defining trace propagation across HTTP headers.

Signed-off-by: Colin James <colin.barr@cloud.com>