CP-51772: Repair xapi-cli-server's tracing
Use trace propagators to endow the Http.Request.t used for XML-RPC
requests with in-service information, in order to propagate it.

As before, the parent span is named "xe <cmdname>" and its span-id is
what makes its way into the request. If the endpoint receiving the
request wishes to, they can derive subsequent in-service tracing, rooted
from this, by using with_tracing as define in Tracing's Propagator.

Signed-off-by: Colin James <colin.barr@cloud.com>