http-svr: change request_of_bio(_exn) to read_request(_exn)
The function is now passed a plain fd rather than a Buf_io.t ("bio")
value, as it does not actually use the buffered channel and just read
from the fd directly (using `Http.read_http_request_header`).

There used to be an older version if `request_from_bio` that had an
option to read requests in a different way and that did use Buf_io. This
was called the "slow path" and was removed in bc2ff459 in favour of the
current "fast path". This further clean-up opportunity was missed at
that time.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>