http-svr: remove read from Buf_io in read_body
At this point, the only function in the entire code base that read from
a Buf_io.t is `Http_svr.read_body` (apart from a test for Buf_io).
However, it only does so if the buffer is not empty and falls back to
reading directly from the fd is not. And since nothing else reads from a
Buf_io, the buffer is always empty...

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>