xapi_http: unify cases in add_handler
The main difference between the BufIO and FdIO cases was that the former
calls `assert_credentials_ok` with the `callback` in its `~fn`
parameter, while the latter executed the `callback` directly after the
credentials check.

The function `assert_credentials_ok` either calls `fn` or raises an
exception. Well, nearly... It did not actually call `fn` in the unix
socket case, where checks are bypassed. This looks unintended and this
patch corrects it. This only affects the following handlers in xapi,
which use BufIO and require RBAC checks: post_remote_db_access,
post_remote_db_access_v2, get_wlb_report, get_wlb_diagnostics,
get_audit_log. I guess those were simply never used on the unix socket.

The other thing that happens when using `~fn` is that the
function `Rbac_audit.allowed_post_fn_ok` is called after `~fn`. This
writes an "ALLOWED OK" line to the audit log. I don't see a reason not
to do the same in all cases.

The outcome is that now both cases of `add_handler` do the same and only
the channel types are different. In the following commit the two handler
types are joining into a single one, which is now easier.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>