CA-391651 - rrd: Carry indices with datasources
ds_update* functions previously relied on being called for the entirety of
datasources of a certain RRD. Instead, operate on chunks of datasources
provided by the same plugin. These chunks are not contiguous (more detailed
explanation in the comments in the code), so indices to rrd_dss and the
associated structures need to be carried with each datasource.

Also turns the value*transform tuple into a type for more explicit accesses.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>