Remove use of deprecated syslog Standard* type
Newer systemd warns that the "syslog" StandardOutput/StandardError type
is deprecated and automatically uses the journal instead. Fix this by
removing the explicit setting of StandardOutput/StandardError. Instead,
the service will use the default values configured in systemd
(DefaultStandardOutput/DefaultStandardError) which for a XenServer
system will result in the output going to rsyslog.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>