CA-400060: Introduce new field for sm class
`host_pending_features` represents the features that are available on
some of the hosts in the pool, but not all of them. Note the way this
field is initialised in the `SM.create` code means that it will only
contain new features that appear during upgrade. This means a feature
that is added into `SM.features` during creation will stay there even if
it is not available on all hosts. But we should not end up in this
situation in the first place.

Also change the meaning of `Sm.features` to be pool-wide.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>