CP-51489: Classify threads based on http requests.
For now, the thread executing `Xapi.server_init` and it's children are
classified as External. The only excception are http requests that come
through the smapi internally. If those contain the originator header
with the value set as "sm", the thread executing the request will be
classified as internal.

This represents the first phase of classifing xapi threads as internal
vs external.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>