CP-50537: Propagate originator as a http request header
XenAPI.py now passes an additional originator header when making
requests to xapi, if the "ORIGINATOR" env var is present.
Sm_exec now passes an env var, "ORIGINATOR=SM".

To classify the threads correctly, we first need to determine the
requests originators. This commit makes it possibly to explicitly
state the originators as headers.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>