CA-391651: rrdd_monitor - Handle missing datasources by resetting them explicitly
If RRD's datasource disappears and is not reported by the plugin (or collected
by xcp-rrdd), then it's not going to be refreshed, keeping its old value and
timestamp. Previously, this was handled nicely because the whole RRD was
updated at once, and updates were defaulted to VT_Unknown, Identity before
being assigned actually collected values. After RRD collection was changed to
process datasources in chunks per source, we need to explicitly find the ones
that weren't updated on this iteration, and explicitly reset them.

(This is not just a theoretical exercise, it can happen when a VIF gets
unplugged or destroyed, for example, with its stats not being reset and
continuing to display old values)

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>