CP-50537: Always reset `_extra_headers` when making a connection.
Clears the `extra_headers` of `UDSTransport` instance when making a
connection.

Previously, this was done only when tracing was enabled inside the
`with_tracecontext` method to avoid the header duplication when
`make_connection` was used multiple times. Currently, there is not
other use of `add_extra_headers` or other update to the
`_extra_headers`, making it safe to clear it when we make a new
connection.

(`xmlrpclib.Transport` updates the `_extra_headers` attribute only
inside `make_connection` method but we override this method with our
own for `UDSTransport`.)

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>