CP-45016: Add support for specifying nbd export in sparse_dd
Previously we hardcoded the export to be empty string when the nbd
client negotiates with the server. This patch allows specifying this
as a protocol parameter.

This is useful for SXM of SMAPIv3 SRs when we copy the snapshot from
source to destination, where the export name is needed when we copy
using nbd.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>