Remove notion of weird string from sexpr library
Removes the remnants of an escaping mechanism once used in xapi. It
appears that the format "<<xxx<value<xxx<" was an early escaping format.

Also removes the effort at line tracking from the lexer actions. Since
the .mli for the generated lexer was only present to have this line
variable, it is deleted.

Signed-off-by: Colin James <colin.barr@cloud.com>