CP-52039: Drop Semaphore from Xapi_stdext_threads
Prior to version 4.12, OCaml's standard threads library (systhreads) had
no builtin concept of a semaphore, so one was implemented in
Xapi_stdext_threads.

We replace all usages of this with Semaphore.Counting from the standard
library and remove the implementation from Xapi_stdext_threads.

Technically, the interface provided by the previous semaphore is more
general: it permits arbitrary adjustments to the semaphore's counter,
allowing for a "weighted" style of locking. However, this is only used
in one place (with a weight value of 1, which is the same
decrement/increment value as normal).

Signed-off-by: Colin James <colin.barr@cloud.com>