CA-401242: avoid long-running, idle connections on VDI.pool_migrate, backport
When a VDI.pool_migrate starts at a pool member, a connection between the
coordinator and that host remains open for the duration of the migration. This
connection is completely idle. If the migration lasts for more than 12 hours,
stunnel closes the connection due to inactivity, which cancels the migration.

To avoid this use an internal API that uses short-running connection whenever
possible to avoid interrupting the migration.

merged cherry-pick of 0149ee5c2433b8995d8b75f28c2037c9da4cba80

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>