Do not include xapi-clusterd.service in toolstack.target
This service is not enabled and started by default, but used on-demand
whenever it is needed for clustering. The current Wants= option in the
toolstack.target is causing xapi-clusterd.service to be started by
xe-toolstack-restart even if it is not enabled.

The fix is to replace Wants= in toolstack.target with WantedBy= in
xapi-clusterd.service, as the latter only installs the dependency when
enabling the service.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>