CA-401324: Update pvsproxy socket location
A previous commit changed the socket location to /run/pvsproxy but this
is problematic because the pvsproxy daemon runs as a deprivileged user
and cannot create the socket.

Instead, update the path to a location that the daemon has permission to
create. Add a fallback to the original path to cope with older pvsproxy
daemons. This fallback can be removed in the future.

Co-developed-by: Pau Ruiz Safont <Pau.ruizsafont@cloud.com>
Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>