CP-49133: Use a worker pool for handling requests
Instead of creating a new thread for each request, enqueue tasks for a
pool of dedicated worker threads to process. To avoid delaying (and
deadlocking, in the case of a task that calls back into xapi), we
fallback to spawning a new thread if the task queue cannot be accessed
immediately. In practice, we found that under a reasonable load, every
request was handled by the pool (of size 16).

You can specify an integer [0-16] for the worker pool size via the
"request-worker-pool-size" option in xapi.conf.

Signed-off-by: Colin James <colin.barr@cloud.com>