CP-50537: Add a guard in `xapi_globs`, `Xapi_globs.tgroups_enabled`.
Adds a configurable variable in `xapi_globs`, `tgroups_enabled` that is
meant to ask a guard for tgroup classification of the threads. If the
guard is `false` all Tgroups functionality should act as a no op.

For instance, adding the line:

tgroups-enabled = false

will result in the thread classification being skipped.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>