CP-45016: Delay VDI.compose in SXM
`VDI.compose` operation links a child volume to a parent, which changes
the internal structure of the volume. This is unsafe to do while the
mirroring datapath is activated. Instead, we move `VDI.compose` into
`receive_finalize`, which is when the source VM is paused, and the
mirroring datapath deactivated.

Implement `receive_finalize2` for this purpose, and also maintain
backwards compatibility.

This change also makes the dummy snapshot of the mirror VDI on the
destination host redundant, as the GC only will start if there is a
parent with a single child in the VHD tree, which will not happen if
compose is not called. Since compose is only called when mirroring is
finished, we do not need to prevent GC from interfering the mirroring
operation.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>