Remove possible systematic leak in Imperative priority queue
The queue is implemented with an Array larger than the content.
Instead of copying an initial value taken from first insert or potentially later
when the Array is expanded provide a "default" value to use as a filler.
This allows to provide a value not having references to external object so
not extending their lifetime in an unpredictable way.

Signed-off-by: Frediano Ziglio <frediano.ziglio@cloud.com>