CA-388564: move qemu-dm to vm.slice
Qemu moves itself to the root cgroup, which makes it a sibling of control.slice.
All toolstack processes are in control.slice, and even though control.slice gets 10x share of CPU,
if we have 1000 qemu processes there is still a big imbalance (toolstack gets 10/1010 share of the CPU)

Qemu-wrapper already has the ability to move the process to another cgroup, so use it.
tapdisk and qemu-datapath are already in vm.slice, so move qemu-dm there too.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>