CP-45016: Implement a new nbd proxy handler
The new http handler nbd_proxy accepts a connection from an nbd client,
and acts as a proxy between the client and the nbd server of the
underlying storage backend, as returned by get_nbd_server.

This proxy is used in place of the old style of fd passing to tapdisks
for VDI copying during storage migration, since the new nbd server from
qemu-dp does not support accepting fds.

This might also be used in the future for outbound storage migration for
mirroring.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>