Revert "XSI-1706/CP-51295: limit open TCP connections to 250 and increase socket backlog"
This reverts commit af7b9c441158e6f4d5cbaf8da5e1879f0cbe51fd.

We found that the lower limits are too low for large pools and reverting
this. We understand that this risks running out of file descriptors,
which was the motivation to lower the limits and to introduce rate
limits. The problem does not exist in XS8 and we hope to manage this
problem until the end of support of this release.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>