CA-403620: Drop the usage of fuser in stunnel client proxy
The drawback of fuser is that it gets too many things involved. E.g. it
is observed that it got stuck on cifs kernel module.

This change uses a cleaner way to remember the stunnel client proxy.
Even when the xapi restarted unexpectedly, it can stop the remnant
stunnel proxy and start a new one.

Signed-off-by: Ming Lu <ming.lu@cloud.com>