CA-404013: do not relock the mutex when backing up rrds
The point of using try_lock is to not get the thread suspended while trying to
hold the mutex. Releasing it and calling `lock` may suspend the thread and
defeats the purpose of using try_lock in the first place.

Reorganise the sequence to read and copy all the rrds first while under the
locked mutex, release it, and then archive the copies.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>