CA-399260: Keep both new and old certs during the switchover
In host-refresh-server-certificate, host generates and applies
new pool certificate after stunnel restart. There is 5s that
other hosts don't trust the new certificate. Then operations
related with xapi:pool SNI tls connection will fail. Both
the old and new certificates shall be trusted during the
switchover to avoid this. At the end of the refresh procedure,
remove_stale_cert will rename the new pem to pem and update
ca bundle.
So, both pem and new pem shall be filled into bundle file in
update-ca-bundle.sh. As the new pem is a temp file and has a
risk to be deleted by other instance between find and cat.
Ignore the cat new pem fail in the script. If the new pem
doesn't exist at the moment, just skip it and go on.

Signed-off-by: Changlei Li <changlei.li@cloud.com>